msm: pm: Fix errors with gcc 4.9
Switching to gcc 4.9 throws a "iteration 3u undefined behavior" when
variables are updated by function returns and the function in itself
takes the variable as a parameter. Fix issue by using a local variable
to store the intermediate value

Also, fix a issue where the variable counter_name could be indexed
incorrectly by explicitly declaring the size of the variable.

Change-Id: I5e0adfef36b139ee2173a4242193cabd3413911e
Signed-off-by: Mahesh Sivasubramanian <msivasub@codeaurora.org>